Things seem to be working again after some big refactoring
of the module system. Just need to take out one million debugging calls.

This part of the system has become a bit of a tangled mess, partly because we're
already trying to keep AOT in mind, but it's likely that when we do AOT this
will all change significantly again...

But at least we have packages and modules now with full QSyms as names.