Bunch of improvements around package/module resolution
Say you do (:import [jsdom :from "jsdom"]), we need to resolve this to the
`node_modules` inside the piglet package where this is being called from, but
node resolves this based on the file that the `import` statement is coming from,
i.e. piglet/lang.mjs. So we take this resolution in our own hands, mimicking
node's logic.

Also improve PDP to include package and lcation in each message, so we can
correctly resolve things.

Move the verbosity level from the level into the compiler, so we can do `-v` and
get the compiled JS source, also when using PDP.