Make js:Object Seqable
We want interop to be smooth, and plain Array objects are also Seqable.

Not 100% sure that this is a good idea, since it means that any object that
isn't explicitly Seqable now gets an implement that loops over its entries,
which for constructed objects often won't make sense.

Still, I think this is fundamental enough to support, while other things like
implementing DictLike for Object are less likely to happen.