Fix destructuring in protocol implementations
- Why did it break?

extend-type/extend-protocol need to know the arities of the function its given,
for this it checks function.length. However since the recent destructuring
overhaul, we emit zero-arg functions for any function that has destructuring,
instead doing sequential destructuring on the `arguments` magical variable.

- How we fixed it

The analyzer will now set `:arities` metadata on functions, which Protocol will
check, before falling back to fn.length.

This also leads the way to having functions with multiple arities, in which case
we want the analyzer to be able to inspect the metadata to generate optimized
dispatch code. It will also allow us to do arity checking at compile time.